feat(combat): effective skill for percent recipes, and tactics out of the math
Three foundation changes of the 15.25 combat model, in one commit because they
are the same player's numbers and the tests read across them.

1. A percent skill recipe scales from the TOTAL skill

Blood Rage, Protector, Sharpshooter and the other percent stances computed their
bonus from Player::getBaseSkill - the raw trained skill, equipment excluded. The
update says they scale from the total. Reading getSkillLevel instead would be a
self-referential loop: it already contains the flat value the recipe itself put
on the player, so every re-derivation would compound (100 -> 130 -> 169 -> 219).

Player now tracks varSkillsFromPercent alongside varSkills - exactly the part of
varSkills that percent recipes contributed. getSkillLevelForPercentScaling runs
the same aggregation as getSkillLevel with that part subtracted, so the source a
recipe reads can never contain a recipe's output, its own or another's. Counted
as source: trained skill, loyalty, equipment, imbuements, flat condition
recipes, Wheel stats, Weapon Proficiency skill bonuses. Excluded: every
percent-derived value.

ConditionAttributes::reapplyPercentSkills replaces updatePercentSkills. It moves
the difference between what the recipe should have on the player and what it
does have (skillsFromPercentApplied), so applying it again is a no-op. It runs
on startCondition, on a merge, on endCondition, and from
Player::refreshPercentSkillRecipes whenever a non-percent source changes -
setVarSkill and a skill advance both call it. Two recipes on one skill read the
same source, so they add and their order does not matter.

The blob now saves the recipe and not the value it produced: skills[] holds the
flat recipe only. CONDITIONATTR_PERCENT_RECIPES_SEPARATE marks the new format;
a blob without it has its flat value dropped for any skill that also carries a
percentage, or every existing player with a stance would log in with the bonus
counted twice.

2. Attack / Balanced / Defense no longer weigh anything

Player::usesModernCombatModel() is the single predicate: the client's protocol
profile has TacticsWithoutFightMode (the current 15.25 profile does; Tibia 11.00
and 8.60 do not), and a player with no client is on the modern model. On it,
getAttackFactor and getDefenseFactor are 1.0, attackTotal counts the equipment
as it is, getDefense's zero-skill branch returns 1, and
PlayerWheel::calculateMitigation has no fight factor. A legacy client keeps
every pre-15.25 number, including raw item values; the two models never mix on
one player.

FightMode_t, its protocol bytes and its Lua enum are untouched - this is the
mathematics only. Player::getCombatTacticsMitigation, which had no caller, is
deleted.

3. Weapon Attack +20%, shield Defence +30%, spellbook Defence +60%

src/creatures/combat/effective_combat_values.hpp holds the three percentages and
the classification that picks between them. Raw item data stays canonical -
Item::getAttack and getDefense still answer what items.xml says, which is what
the market, item descriptions and serialisation read. The Player exposes the
effective values (getEffectiveWeaponAttackValue, getEffectiveOffhandDefense,
getEffectiveShieldDefense) and every consumer takes them from there, so each
percentage lands exactly once: the weapon damage formulas, getDefense,
getDefenseEquipment, calculateMitigation and both client stat payloads.

The values are doubles so the fraction survives to whatever formula consumes it;
where an integer is needed first, toInteger rounds half away from zero and is
the only rounding point. Not compensated: the bare-fist constant, the
proficiency's flat ATTACK_DAMAGE (its own layer, added after), spells and runes,
and any off-hand item that is neither a shield nor a spellbook. The elemental
share the client shows stays a ratio of the raw pair, where the compensation
cancels out.

Player:getEquippedShieldDefense in Lua delegates to the new
Player:getEffectiveShieldDefense, so Shield Bash and Shield Slam scale from the
same number as normal shield combat and no script applies the 30% again. Their
damage formula is still the unverified inference PR #49 marked.

Tests: 29 cases across three new files - the percentage arithmetic and
classification, the effective-skill matrix (equipment counts, five refreshes do
not compound, relog restores once, two recipes add, a pre-separation blob drops
its stale value), and the tactics matrix (every value identical in all three
fight modes, a legacy client still different, each compensation once, raw data
untouched).

+++ src/creatures/combat/combat.cpp
@@ -1910,7 +1910,7 @@ void ValueCallback::getMinMaxValues(const std::shared_ptr<Player> &player, Comba
 	LuaScriptInterface::setMetatable(L, -1, "Player");
 
 	int16_t elementAttack = 0; // To calculate elemental damage after executing spell script and get real damage.
-	int32_t attackValue = 7; // default start attack value
+	double attackValue = 7; // default start attack value, effective (15.25 compensated) once a weapon fills it
 	int parameters = 1;
 	bool shouldCalculateSecondaryDamage = false;
